test(task-migrate-clean-batch): 迁 user_defaults 31 用例到 gui_unit_test，并补上该 target 的个人默认值隔离
420.5 Step 6，Group A 收官（M1）。本批单文件迁移量最大的一个。

**迁移**
- 新增 test/unit-correctness/gui/test_user_defaults.cpp（31 个 TEST）
- 删除 test/gui/functional/test_gui_user_defaults.cpp + 注册声明/调用点
- 去掉 test_gui_shared.hpp 依赖，显式 include gui/file_io.hpp
  （SerializeGuiStateJson / DeserializeGuiStateJson 原先由它间接带入）
- gui_unit_test 补 LUMICE_GUI_APP_CPP_PATH 编译定义：
  bg_degrade_is_wired_into_every_document_path 是读 app.cpp 源码的结构性闸

**顺带修一个真缺陷：gui_unit_test 此前没有个人默认值隔离基线**
switch_gui_test_baseline_is_isolated 断言的是"进程基线下，磁盘上的 override
不该被读到"。gui_test 由自己的 main() 装 kTestHarnessUserConfigDefault；
gui_unit_test 与非 GUI target 共用 test_main.cpp（那边碰不到 lumice_gui_obj），
于是进程源一直停在 kAutoDetect —— 即 user_defaults.cpp 里写明的"未设置值"，
会去读开发机真实的 OS 配置目录。

实测：把 HOME 指向一个存有 bg_alpha 个人默认值的目录，该用例在 fresh 进程里
直接变红。这不是这一个用例的问题——本 binary 里今后任何走无参
MakeNewDocumentState() 的用例都同样暴露，所以修在 harness 而不是修在用例。

- 新增 test/unit-correctness/gui/gui_unit_test_env.cpp：gtest 全局 Environment，
  在任何用例前装上与 gui_test 相同的 kTestHarnessUserConfigDefault
- 该用例更名 switch_gui_test_baseline_is_isolated →
  switch_harness_baseline_is_isolated（原名指向的是另一个 binary），注释同步

核验：断言 193→193（IM_CHECK 106→EXPECT_TRUE、IM_CHECK_EQ 87→EXPECT_EQ）；
gui_test 静态用例数 541→510（= M1 目标）；gui_unit_test 34→65 全绿，
且在敌意 HOME（存有个人默认值）下隔离跑与整套跑均 65/65 绿。
